shader: Use std::bit_cast (C++20) to convert float<->int
Fixes ` warning: dereferencing type-punned pointer will break
strict-aliasing rules [-Wstrict-aliasing]` when compiling with `-Wall`.
