+++ vita3k/shader/include/shader/usse_constant_table.h
@@ -17,6 +17,8 @@
 
 #pragma once
 
+#include <util/bit_cast.h>
+
 #include <cstdint>
 
 namespace shader::usse {
@@ -24,8 +26,8 @@ namespace shader::usse {
 static const std::uint32_t NaN_raw = 0x7FFF7FFF;
 static const std::uint32_t neg_NaN_raw = 0xFFFFFFFF;
 
-static const float NaN = *reinterpret_cast<const float *>(&NaN_raw);
-static const float neg_NaN = *reinterpret_cast<const float *>(&neg_NaN_raw);
+static const float NaN = std::bit_cast<float>(NaN_raw);
+static const float neg_NaN = std::bit_cast<float>(neg_NaN_raw);
 
 constexpr std::uint32_t f32_constant_table_bank_0_raw[] = {
     0x00000000,
